Refactor CreateWallet onboarding
