+++ packages/extension/src/components/pages/CreateWallet/ConfirmSeed/index.ts
@@ -1,0 +1,1 @@
+export * from './ConfirmSeed';
