+++ packages/extension/src/components/pages/CreateWallet/CreatePassword/index.ts
@@ -1,0 +1,1 @@
+export * from './CreatePassword';
