+++ packages/extension/src/components/pages/CreateWallet/SecretSeed/index.ts
@@ -1,0 +1,1 @@
+export * from './SecretSeed';
