+++ packages/extension/src/components/pages/CreateWallet/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './steps';
