+++ packages/extension/src/components/pages/CreateWallet/constants/steps.ts
@@ -1,0 +1,1 @@
+export const STEPS = 4;
